[CLN](foundation-api): Extract init schema
Move the pure schema and embedding-function builders for Foundation
collections (qwen/splade known EFs, CollectionEmbeddingFunctions, and
foundation_collection_schema) plus their tests out of the init handler
into routes/init_schema.rs.

No behavior change; keeps init.rs focused on orchestration and forms the
base of the init tracing + retry stack.
